requirements.txt
pynput>=1.8.1
pystray>=0.19.5
Pillow>=10.0.0
